+++ src/Glue/CamelCase.php
@@ -2,13 +2,22 @@
 
 namespace Jawira\CaseConverter\Glue;
 
-use const MB_CASE_LOWER;
-use const MB_CASE_TITLE;
-
+/**
+ * Class CamelCase
+ *
+ * Outputs string in _Camel case_ format: thisIsCamelCase
+ *
+ * @package Jawira\CaseConverter\Glue
+ */
 class CamelCase extends UppercaseGluer
 {
+    /**
+     * Format detected words in _Camel case_
+     *
+     * @return string
+     */
     public function glue(): string
     {
-        return $this->glueUsingRules(self::DELIMITER, MB_CASE_TITLE, MB_CASE_LOWER);
+        return $this->glueUsingRules(self::DELIMITER, $this->titleCase, $this->lowerCase);
     }
 }