+++ src/Glue/CobolCase.php
@@ -2,12 +2,22 @@
 
 namespace Jawira\CaseConverter\Glue;
 
-use const MB_CASE_UPPER;
-
+/**
+ * Class CobolCase
+ *
+ * Outputs string in _Cobol case_ format: THIS-IS-COBOL-CASE
+ *
+ * @package Jawira\CaseConverter\Glue
+ */
 class CobolCase extends DashGluer
 {
+    /**
+     * Format detected words in _Cobol case_
+     *
+     * @return string
+     */
     public function glue(): string
     {
-        return $this->glueUsingRules(self::DELIMITER, MB_CASE_UPPER);
+        return $this->glueUsingRules(self::DELIMITER, $this->upperCase);
     }
 }