+++ src/Glue/KebabCase.php
@@ -2,12 +2,22 @@
 
 namespace Jawira\CaseConverter\Glue;
 
-use const MB_CASE_LOWER;
-
+/**
+ * Class KebabCase
+ *
+ * Outputs string in _Cobol case_ format: this-is-kebab-case
+ *
+ * @package Jawira\CaseConverter\Glue
+ */
 class KebabCase extends DashGluer
 {
+    /**
+     * Format detected words in _Kebab case_
+     *
+     * @return string
+     */
     public function glue(): string
     {
-        return $this->glueUsingRules(self::DELIMITER, MB_CASE_LOWER);
+        return $this->glueUsingRules(self::DELIMITER, $this->lowerCase);
     }
 }