+++ src/Glue/MacroCase.php
@@ -2,12 +2,10 @@
 
 namespace Jawira\CaseConverter\Glue;
 
-use const MB_CASE_UPPER;
-
 class MacroCase extends UnderscoreGluer
 {
     public function glue(): string
     {
-        return $this->glueUsingRules(self::DELIMITER, MB_CASE_UPPER);
+        return $this->glueUsingRules(self::DELIMITER, $this->upperCase);
     }
 }