+++ src/Split/SpaceSplitter.php
@@ -8,6 +8,9 @@ class SpaceSplitter extends Splitter
 {
     const PATTERN = SpaceGluer::DELIMITER . '+';
 
+    /**
+     * @return string[]
+     */
     public function split(): array
     {
         return $this->splitUsingPattern($this->inputString, self::PATTERN);