+++ tests/phpunit/CamelCaseTest.php
@@ -3,10 +3,17 @@
 use Jawira\CaseConverter\Glue\CamelCase;
 use PHPUnit\Framework\TestCase;
 
+/**
+ * Class CamelCaseTest
+ */
 class CamelCaseTest extends TestCase
 {
     /**
+     * Testing that `glue()` method is called and `titleCase` and `lowerCase`
+     * properties are being used.
+     *
      * @covers \Jawira\CaseConverter\Glue\CamelCase::glue
+     * @throws \ReflectionException
      */
     public function testGlue()
     {
@@ -16,14 +23,23 @@ class CamelCaseTest extends TestCase
                      ->setMethods(['glueUsingRules'])
                      ->getMock();
 
+        // Setting titleCase and lowerCase properties
+        $reflectionObject  = new ReflectionObject($mock);
+        $titleCaseProperty = $reflectionObject->getProperty('titleCase');
+        $titleCaseProperty->setAccessible(true);
+        $titleCaseProperty->setValue($mock, 123);
+        $lowerCaseProperty = $reflectionObject->getProperty('lowerCase');
+        $lowerCaseProperty->setAccessible(true);
+        $lowerCaseProperty->setValue($mock, 456);
+
         // Configuring stub
         $mock->expects($this->once())
              ->method('glueUsingRules')
-             ->with(CamelCase::DELIMITER, MB_CASE_TITLE)
-             ->willReturn('e1bfd762321e409cee4ac0b6e841963c');
+             ->with(CamelCase::DELIMITER, 123, 456)
+             ->willReturn('dummy-value-32ea');
 
-        /** @var \Jawira\CaseConverter\CamelCase $mock */
+        /** @var \Jawira\CaseConverter\Glue\CamelCase $mock */
         $returned = $mock->glue();
-        $this->assertSame('e1bfd762321e409cee4ac0b6e841963c', $returned, 'Returned value is not the expected');
+        $this->assertSame('dummy-value-32ea', $returned, 'Returned value is not the expected');
     }
 }